fix: transaction sql

+++ app/migrations/3.26.64_add_manager_identification/add_manager_identification.sql
@@ -42,9 +42,29 @@ CREATE TABLE `manager_identification` (
   	CONSTRAINT `manager_identification_ibfk_5` FOREIGN KEY (`school_inep_id_fk`) REFERENCES `school_identification` (`inep_id`) ON DELETE CASCADE ON UPDATE CASCADE
 );
 
--- MIGRANDO OS DADOS DE GESTOR JÁ CADASTRADOS EM SCHOOL_IDENTIFICATION
-BEGIN TRANSACTION
+SET @managers := (SELECT 
+s.inep_id as `school_inep_id_fk`, 
+s.manager_name as `name`,
+s.manager_email as `email`,
+'01/01/1998' as `birthday_date`,
+1 as `sex`,
+3 as `color_race`,
+1 as `nationality`,
+s.manager_role as `role`,
+1 as `residence_zone`,
+s.manager_access_criterion as `access_criterion`,
+s.manager_contract_type as `contract_type`,
+s.manager_cpf as `cpf`,
+s.number_ato `number_ato`,
+0 as `filiation`,
+76 as `edcenso_nation_fk`,
+s.edcenso_uf_fk as `edcenso_uf_fk`,
+s.edcenso_city_fk as `edcenso_city_fk`
+FROM `school_identification` s
+WHERE s.manager_name IS NOT NULL);
 
+-- MIGRANDO OS DADOS DE GESTOR JÁ CADASTRADOS EM SCHOOL_IDENTIFICATION
+START TRANSACTION
 INSERT INTO `manager_identification` (
 	`school_inep_id_fk`, 
 	`name`, 
@@ -64,35 +84,25 @@ INSERT INTO `manager_identification` (
 	`edcenso_uf_fk`,
 	`edcenso_city_fk`
 )
-SELECT 
-s.inep_id as `school_inep_id_fk`, 
-s.manager_name as `name`,
-s.manager_email as `email`,
-'01/01/1998' as `birthday_date`,
-1 as `sex`,
-3 as `color_race`,
-1 as `nationality`,
-s.manager_role as `role`,
-1 as `residence_zone`,
-s.manager_access_criterion as `access_criterion`,
-s.manager_contract_type as `contract_type`,
-s.manager_cpf as `cpf`,
-s.number_ato `number_ato`,
-0 as `filiation`,
-76 as `edcenso_nation_fk`,
-s.edcenso_uf_fk as `edcenso_uf_fk`,
-s.edcenso_city_fk as `edcenso_city_fk`
-FROM `school_identification` s
-WHERE s.manager_name IS NOT NULL;
+@managers;
 
-COMMIT
+-- Conta quantos registros foram afetados na migração
+SELECT ROW_COUNT() INTO @affected_rows;
 
--- DELETANDO COLUNAS DE SCHOOL_IDENTIFICATION
-ALTER TABLE school_identification 
-DROP COLUMN `manager_name`, 
-DROP COLUMN `manager_email`, 
-DROP COLUMN `number_ato`, 
-DROP COLUMN `manager_role`, 
-DROP COLUMN `manager_cpf`, 
-DROP COLUMN `manager_contract_type`, 
-DROP COLUMN `manager_access_criterion`;
+-- Conta quantos registros deveriam ser criados na migração
+SELECT COUNT(*) INTO @count FROM (@managers);
+
+-- Verifica se a migração foi bem-sucedida
+IF @affected_rows = @count THEN
+	ALTER TABLE school_identification 
+	DROP COLUMN `manager_name`, 
+	DROP COLUMN `manager_email`, 
+	DROP COLUMN `number_ato`, 
+	DROP COLUMN `manager_role`, 
+	DROP COLUMN `manager_cpf`, 
+	DROP COLUMN `manager_contract_type`, 
+	DROP COLUMN `manager_access_criterion`;
+	COMMIT; -- Confirma a transação
+ELSE
+    ROLLBACK; -- Desfaz a transação se a migração não foi bem-sucedida
+END IF;
